Anteacher/requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
yfinance
finance-datareader
google-generativeai
python-dotenv
requests
feedparser
transformers
torch
scipy
